requirements.txt
langchain
streamlit
openai
chromadb
tiktoken
python-dotenv